created and updated most of the pages
The following views that currently exist are:
about
entercourses
main
newstudent
registerconfirm
studentview
viewcourses

There are some pages that can't be accessed right now unless they are
manually typed in as a url because the login and register buttons aren't
running right now (ex: registerconfirm). When grunt serve starts running
it goes directly to the log in page, but I made the home page go to the
student view. I put a login button in the top navigation bar as a
temporary thing but in the future it could become a logout button if the
user is logged in?